174: extract auto dismiss logic from toast
